requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
google-genai>=1.0.0
numpy>=1.24.0
